# add optimized serialization methods