Add binary-compatible serialization/deserialization which avoids allocating memory unless absolutely necessary. `ToBytes` is applicable anywhere you need to send multiple digests across the wire, `FromBytes` is a little stranger since it obliterates the existing digest. But our use case can take advantage of this in many cases.

Note rather substantial performance delta between old and new methodologies.

```
BenchmarkAsBytes-4           	  100000	     12925 ns/op	    1960 B/op	     209 allocs/op
BenchmarkToBytes-4           	 2000000	       634 ns/op	       0 B/op	       0 allocs/op
BenchmarkFromBytes-4         	   50000	     32229 ns/op	   17808 B/op	     112 allocs/op
BenchmarkFromBytesMethod-4   	 2000000	       836 ns/op	       0 B/op	       0 allocs/op
```